+++ google-cloud-storage/src/main/java/com/google/cloud/storage/spi/v1/HttpStorageRpc.java
@@ -24,6 +24,7 @@ import com.google.api.client.googleapis.batch.BatchRequest;
 import com.google.api.client.googleapis.batch.json.JsonBatchCallback;
 import com.google.api.client.googleapis.json.GoogleJsonError;
 import com.google.api.client.http.ByteArrayContent;
+import com.google.api.client.http.EmptyContent;
 import com.google.api.client.http.GenericUrl;
 import com.google.api.client.http.HttpHeaders;
 import com.google.api.client.http.HttpRequest;
@@ -31,6 +32,7 @@ import com.google.api.client.http.HttpRequestFactory;
 import com.google.api.client.http.HttpRequestInitializer;
 import com.google.api.client.http.HttpResponse;
 import com.google.api.client.http.HttpResponseException;
+import com.google.api.client.http.HttpStatusCodes;
 import com.google.api.client.http.HttpTransport;
 import com.google.api.client.http.InputStreamContent;
 import com.google.api.client.http.json.JsonHttpContent;
@@ -751,7 +753,8 @@ public class HttpStorageRpc implements StorageRpc {
   public long getCurrentUploadOffset(String uploadId) {
     try {
       GenericUrl url = new GenericUrl(uploadId);
-      HttpRequest httpRequest = storage.getRequestFactory().buildPutRequest(url, null);
+      HttpRequest httpRequest =
+          storage.getRequestFactory().buildPutRequest(url, new EmptyContent());
 
       httpRequest.getHeaders().setContentRange("bytes */*");
       // Turn off automatic redirects.
@@ -763,9 +766,9 @@ public class HttpStorageRpc implements StorageRpc {
       try {
         response = httpRequest.execute();
         int code = response.getStatusCode();
-        String message = response.getStatusMessage();
-        if (code == 201 || code == 200) {
-          throw new StorageException(0, "Resumable upload is already complete.");
+        if (HttpStatusCodes.isSuccess(code)) {
+          // Upload completed successfully
+          return -1;
         }
         StringBuilder sb = new StringBuilder();
         sb.append("Not sure what occurred. Here's debugging information:\n");
@@ -773,14 +776,18 @@ public class HttpStorageRpc implements StorageRpc {
         throw new StorageException(0, sb.toString());
       } catch (HttpResponseException ex) {
         int code = ex.getStatusCode();
-        if (code == 308 && ex.getHeaders().getRange() == null) {
-          // No progress has been made.
-          return 0;
-        } else {
+        if (code == 308) {
+          if (ex.getHeaders().getRange() == null) {
+            // No progress has been made.
+            return 0;
+          }
           // API returns last byte received offset
           String range = ex.getHeaders().getRange();
           // Return next byte offset by adding 1 to last byte received offset
           return Long.parseLong(range.substring(range.indexOf("-") + 1)) + 1;
+        } else {
+          // Something else occurred like a 5xx so translate and throw.
+          throw translate(ex);
         }
       } finally {
         if (response != null) {
@@ -881,38 +888,30 @@ public class HttpStorageRpc implements StorageRpc {
       if (kmsKeyName != null && kmsKeyName.contains("cryptoKeyVersions")) {
         object.setKmsKeyName("");
       }
-      Insert req = storage.objects().insert(object.getBucket(), object);
-      GenericUrl url = req.buildHttpRequest().getUrl();
-      String scheme = url.getScheme();
-      String host = url.getHost();
-      int port = url.getPort();
-      port = port > 0 ? port : url.toURL().getDefaultPort();
-      String path = "/upload" + url.getRawPath();
-      url = new GenericUrl(scheme + "://" + host + ":" + port + path);
+      Insert req =
+          storage
+              .objects()
+              .insert(object.getBucket(), object)
+              .setName(object.getName())
+              .setProjection(Option.PROJECTION.getString(options))
+              .setPredefinedAcl(Option.PREDEFINED_ACL.getString(options))
+              .setIfMetagenerationMatch(Option.IF_METAGENERATION_MATCH.getLong(options))
+              .setIfMetagenerationNotMatch(Option.IF_METAGENERATION_NOT_MATCH.getLong(options))
+              .setIfGenerationMatch(Option.IF_GENERATION_MATCH.getLong(options))
+              .setIfGenerationNotMatch(Option.IF_GENERATION_NOT_MATCH.getLong(options))
+              .setUserProject(Option.USER_PROJECT.getString(options))
+              .setKmsKeyName(Option.KMS_KEY_NAME.getString(options));
+      GenericUrl url = req.buildHttpRequestUrl();
+      url.setRawPath("/upload" + url.getRawPath());
       url.set("uploadType", "resumable");
-      url.set("name", object.getName());
-      for (Option option : options.keySet()) {
-        Object content = option.get(options);
-        if (content != null) {
-          url.set(option.value(), content.toString());
-        }
-      }
+
       JsonFactory jsonFactory = storage.getJsonFactory();
       HttpRequestFactory requestFactory = storage.getRequestFactory();
       HttpRequest httpRequest =
           requestFactory.buildPostRequest(url, new JsonHttpContent(jsonFactory, object));
       HttpHeaders requestHeaders = httpRequest.getHeaders();
       requestHeaders.set("X-Upload-Content-Type", detectContentType(object, options));
-      String key = Option.CUSTOMER_SUPPLIED_KEY.getString(options);
-      if (key != null) {
-        BaseEncoding base64 = BaseEncoding.base64();
-        HashFunction hashFunction = Hashing.sha256();
-        requestHeaders.set("x-goog-encryption-algorithm", "AES256");
-        requestHeaders.set("x-goog-encryption-key", key);
-        requestHeaders.set(
-            "x-goog-encryption-key-sha256",
-            base64.encode(hashFunction.hashBytes(base64.decode(key)).asBytes()));
-      }
+      setEncryptionHeaders(requestHeaders, "x-goog-encryption-", options);
       HttpResponse response = httpRequest.execute();
       if (response.getStatusCode() != 200) {
         GoogleJsonError error = new GoogleJsonError();